feat: added a separate decoder module
Auth jwt token is being verified by the api gateway but is not manipulated in any way. Gateway passes the whole token to the backend in Authorization header if the request is authorized. If the backend needs some user data form the token, it should decode it itself. That is what this module does