Lobby functionality over WebSocket (#5)
* feat: added a basic ws gateway example

* refactor: rewrote gateway to use express-ws

* feat: added global token validation

* fix: changed lobby to use username instead of connection id

* refactor: use express-ws in lobby service

used router to make a new ws endpoint so that all endpoints are served on the same server on the same port

* feat: added joining and leaving functionality to lobby via ws

* fix: some minor fixes and cleanup

* fix: added readme to be finished later

* refactor: gateway refactor

Split up the ws gateway into several modules for the sake of readability and responsibility dispersion. Performed some pruning and generalization of functions

* fix: gateway refactor followup fixes

* feat: added chat to lobby and somewhat pruned the code

* fix: added validation for messages from backend

* fix: fixed a bug with multi bearer denial

Fixed a bug when in case a client uses the same token to connect to the same endpoint, the backend receives events in the wrong order. Also if this was the case, the messages from the backend were not sent back to the new client.

* fix: moved ws gw to a new port

* fix: added gw and lobby descriptions

* fix: removed lobby http join and leave stubs, aded cosmetic changes to docs

Co-authored-by: Nestor Sapuntsov <nestor.sapuntsov@Nestors-MacBook-Pro.local>
Co-authored-by: Nestor Sapuntsov <nestor.sapuntsov@nestors-macbook-pro.home>