fix: changed auth approach
/login and /register endpoints now send signed token in their responses.
Service no longer has endpoint for fetching key(s). Instead the key must be passed as an env.
Service now has a new endpoint /validate that validates a token and either returns 200 with it's payload in the body, or 400 if the token is invalid. This endpoint will be used by the new gateway in the future.