Gateway refactor (#7)
* fix: changed auth approach

/login and /register endpoints now send signed token in their responses.
Service no longer has endpoint for fetching key(s). Instead the key must be passed as an env.
Service now has a new endpoint /validate that validates a token and either returns 200 with it's payload in the body, or 400 if the token is invalid. This endpoint will be used by the new gateway in the future.

* feat: added a module for verifying the token by requesting auth service

* fix: switched to the new validator

* feat: added http endpoints for proxying requests to backend to the ws gateway

* fix: adapted the lobby service to the new gateway

* fix: getting rid of KrakenD

* fix: fixed an issue with passing request body

* fix: cleaned up hosts, ports and endpoints

* fix: fixed cors and updated swagger

* fix: updated readmes and architecture

* fix: minor followup fixes and a readme update

Co-authored-by: Nestor Sapuntsov <nestor.sapuntsov@Nestors-MacBook-Pro.local>